backend/requirements.txt
# ========================================================
# AI Resume Screening System — Backend Dependencies
# ========================================================

# --- FastAPI & Server ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9      # File upload support

# --- Database (PostgreSQL + pgvector) ---
sqlalchemy==2.0.30
asyncpg==0.29.0              # Async PostgreSQL driver
alembic==1.13.1              # Database migrations
pgvector==0.2.5              # pgvector SQLAlchemy type

# --- Pydantic (data validation) ---
pydantic==2.7.1
pydantic-settings==2.3.0     # Settings from .env

# --- Gemini LLM ---
google-generativeai==0.7.2

# --- Resume Parsing ---
pypdf2==3.0.1                # PDF parsing
python-docx==1.1.0           # DOCX parsing

# --- Utilities ---
python-jose[cryptography]==3.3.0  # JWT (future auth)
passlib[bcrypt]==1.7.4            # Password hashing (future auth)
python-dotenv==1.0.1
structlog==24.1.0            # Structured logging
tenacity==8.3.0              # Retry logic

# --- Testing ---
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0                # Async test client
anyio[asyncio]==4.4.0
aiosqlite==0.20.0
